test,ui: move comment error tests to error; remove unused ComposeView
