+++ .gitignore
@@ -19,3 +19,6 @@ app/release/
 bin/
 .vscode/
 *.code-workspace
+
+# logs
+*.log
